Split up model and GQL files
