+++ api/src/models/hardware_spec.rs
