+++ api/src/models/mod.rs
@@ -1,4 +1,4 @@
-//! All Diesel models live here.
+//! All Diesel models live here. One file per DB table.
 //!
 //! Some naming conventions for model structs:
 //! - `Foo`: The "default" model, which contains all fields that in the DB
@@ -8,10 +8,12 @@
 //! - `NewFoo`: A model that can be constructed in our Rust code, and inserted
 //!   into the DB. It should generally define a `.insert()` helper.
 
-mod hardware;
-mod program;
+mod hardware_spec;
+mod program_spec;
 mod user;
+mod user_program;
 
-pub use hardware::*;
-pub use program::*;
+pub use hardware_spec::*;
+pub use program_spec::*;
 pub use user::*;
+pub use user_program::*;
